requirements.txt
# Core data & computation
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Data sources
yfinance>=0.2.36
requests>=2.31.0
requests-cache>=1.1.1

# Feature engineering / ML
scikit-learn>=1.3.0
xgboost>=2.0.0

# Deep learning
torch>=2.1.0
torchvision>=0.16.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Data formats
pyarrow>=13.0.0
fastparquet>=2023.10.1

# CLI
click>=8.1.0
rich>=13.6.0
pyyaml>=6.0.1

# Statistical
statsmodels>=0.14.0

# SHAP explainability (optional but included)
shap>=0.43.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# HTML parsing (for Wikipedia S&P 500 scrape)
lxml>=4.9.0
html5lib>=1.1

joblib>=1.3.0
python-dateutil>=2.8.2
